+++ cpp/Compress/sprintz_xff.cpp
@@ -1553,7 +1553,7 @@ SPRINTZ_FORCE_INLINE int64_t decompress8b_rowmajor_xff_rle(
 
 int64_t decompress8b_rowmajor_xff_rle(const int8_t* src, uint8_t* dest) {
     uint16_t ndims;
-    uint64_t ngroups;
+    uint32_t ngroups;
     uint16_t remaining_len;
     src += read_metadata_rle(src, &ndims, &ngroups, &remaining_len);
     return decompress8b_rowmajor_xff_rle(
